% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.r
\name{count}
\alias{count}
\title{Title}
\usage{
count(InData, element)
}
\arguments{
\item{InData}{Input a vector}

\item{element}{an element to be found}
}
\value{
element happened times in the InData,and the percentage
}
\description{
to find how many times element happened in a vector
}
\examples{
   a=c(1,2,3,4,6,7)
   b=2
   count(a,b)
}

