% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpieces.r
\name{getpieces}
\alias{getpieces}
\title{Title}
\usage{
getpieces(InData, distance)
}
\arguments{
\item{InData}{Input a vector}

\item{distance}{distance between one by one}
}
\description{
get required interval distance pieces
}
\examples{
   a<-c(2,3,5,6,7,8,8,3,4,6,6,1,2,4,56,2,4,5,5)
   distance=2
   getpieces(a,distance)
}

